refactor: split extension methods into separate classes
