+++ src/Commands/ClearCommandGroup.cs
@@ -2,6 +2,7 @@ using System.ComponentModel;
 using System.Text;
 using JetBrains.Annotations;
 using Octobot.Data;
+using Octobot.Extensions;
 using Octobot.Services;
 using Remora.Commands.Attributes;
 using Remora.Commands.Groups;
