+++ src/Responders/MessageDeletedResponder.cs
@@ -1,6 +1,7 @@
 using System.Text;
 using JetBrains.Annotations;
 using Octobot.Data;
+using Octobot.Extensions;
 using Octobot.Services;
 using Remora.Discord.API.Abstractions.Gateway.Events;
 using Remora.Discord.API.Abstractions.Objects;
